# uadk: simplify msg_pool via removing ctx_id