--- a/include/drv/wd_cipher_drv.h
+++ b/include/drv/wd_cipher_drv.h
@@ -50,7 +50,7 @@ struct wd_cipher_msg {
 	__u8 *out;
 };
 
-struct wd_cipher_msg *wd_cipher_get_msg(__u32 idx, __u32 tag);
+struct wd_cipher_msg *wd_cipher_get_msg(__u32 tag);
 
 #ifdef __cplusplus
 }