--- a/include/drv/wd_ecc_drv.h
+++ b/include/drv/wd_ecc_drv.h
@@ -175,7 +175,7 @@ struct wd_ecc_out {
 	char data[];
 };
 
-struct wd_ecc_msg *wd_ecc_get_msg(__u32 idx, __u32 tag);
+struct wd_ecc_msg *wd_ecc_get_msg(__u32 tag);
 
 #ifdef __cplusplus
 }