--- a/wd_aead.c
+++ b/wd_aead.c
@@ -42,7 +42,7 @@ struct wd_aead_setting {
 	struct wd_ctx_config_internal config;
 	struct wd_sched sched;
 	struct wd_alg_driver *driver;
-	struct wd_async_msg_pool pool;
+	struct msg_pool pool;
 	void *dlhandle;
 	void *dlh_list;
 } wd_aead_setting;
@@ -771,8 +771,7 @@ int wd_do_aead_async(handle_t h_sess, struct wd_aead_req *req)
 
 	ctx = config->ctxs + idx;
 
-	msg_id = wd_get_msg_from_pool(&wd_aead_setting.pool,
-				     idx, (void **)&msg);
+	msg_id = wd_get_msg_from_pool(&wd_aead_setting.pool, (void **)&msg);
 	if (unlikely(msg_id < 0)) {
 		WD_ERR("failed to get msg from pool!\n");
 		return -WD_EBUSY;
@@ -797,13 +796,13 @@ int wd_do_aead_async(handle_t h_sess, struct wd_aead_req *req)
 	return 0;
 
 fail_with_msg:
-	wd_put_msg_to_pool(&wd_aead_setting.pool, idx, msg->tag);
+	wd_put_msg_to_pool(&wd_aead_setting.pool, msg->tag);
 	return ret;
 }
 
 struct wd_aead_msg *wd_aead_get_msg(__u32 idx, __u32 tag)
 {
-	return wd_find_msg_in_pool(&wd_aead_setting.pool, idx, tag);
+	return wd_find_msg_in_pool(&wd_aead_setting.pool, tag);
 }
 
 int wd_aead_poll_ctx(__u32 idx, __u32 expt, __u32 *count)
@@ -839,8 +838,7 @@ int wd_aead_poll_ctx(__u32 idx, __u32 expt, __u32 *count)
 		}
 
 		recv_count++;
-		msg = wd_find_msg_in_pool(&wd_aead_setting.pool,
-					    idx, resp_msg.tag);
+		msg = wd_find_msg_in_pool(&wd_aead_setting.pool, resp_msg.tag);
 		if (!msg) {
 			WD_ERR("failed to get msg from pool!\n");
 			return -WD_EINVAL;
@@ -850,8 +848,7 @@ int wd_aead_poll_ctx(__u32 idx, __u32 expt, __u32 *count)
 		msg->req.state = resp_msg.result;
 		req = &msg->req;
 		req->cb(req, req->cb_param);
-		wd_put_msg_to_pool(&wd_aead_setting.pool,
-					       idx, resp_msg.tag);
+		wd_put_msg_to_pool(&wd_aead_setting.pool, resp_msg.tag);
 		*count = recv_count;
 	} while (--tmp);