--- a/wd_comp.c
+++ b/wd_comp.c
@@ -45,7 +45,7 @@ struct wd_comp_setting {
 	enum wd_status status;
 	struct wd_ctx_config_internal config;
 	struct wd_sched sched;
-	struct wd_async_msg_pool pool;
+	struct msg_pool pool;
 	struct wd_alg_driver *driver;
 	void *dlhandle;
 	void *dlh_list;
@@ -324,9 +324,9 @@ void wd_comp_uninit2(void)
 	wd_alg_clear_init(&wd_comp_setting.status);
 }
 
-struct wd_comp_msg *wd_comp_get_msg(__u32 idx, __u32 tag)
+struct wd_comp_msg *wd_comp_get_msg(__u32 tag)
 {
-	return wd_find_msg_in_pool(&wd_comp_setting.pool, idx, tag);
+	return wd_find_msg_in_pool(&wd_comp_setting.pool, tag);
 }
 
 int wd_comp_poll_ctx(__u32 idx, __u32 expt, __u32 *count)
@@ -363,8 +363,7 @@ int wd_comp_poll_ctx(__u32 idx, __u32 expt, __u32 *count)
 
 		recv_count++;
 
-		msg = wd_find_msg_in_pool(&wd_comp_setting.pool, idx,
-					  resp_msg.tag);
+		msg = wd_find_msg_in_pool(&wd_comp_setting.pool, resp_msg.tag);
 		if (unlikely(!msg)) {
 			WD_ERR("failed to get msg from pool!\n");
 			return -WD_EINVAL;
@@ -376,7 +375,7 @@ int wd_comp_poll_ctx(__u32 idx, __u32 expt, __u32 *count)
 		req->cb(req, req->cb_param);
 
 		/* free msg cache to msg_pool */
-		wd_put_msg_to_pool(&wd_comp_setting.pool, idx, resp_msg.tag);
+		wd_put_msg_to_pool(&wd_comp_setting.pool, resp_msg.tag);
 		*count = recv_count;
 	} while (--tmp);
 
@@ -834,7 +833,7 @@ int wd_do_comp_async(handle_t h_sess, struct wd_comp_req *req)
 
 	ctx = config->ctxs + idx;
 
-	tag = wd_get_msg_from_pool(&wd_comp_setting.pool, idx, (void **)&msg);
+	tag = wd_get_msg_from_pool(&wd_comp_setting.pool, (void **)&msg);
 	if (unlikely(tag < 0)) {
 		WD_ERR("failed to get msg from pool!\n");
 		return -WD_EBUSY;
@@ -857,7 +856,7 @@ int wd_do_comp_async(handle_t h_sess, struct wd_comp_req *req)
 	return 0;
 
 fail_with_msg:
-	wd_put_msg_to_pool(&wd_comp_setting.pool, idx, msg->tag);
+	wd_put_msg_to_pool(&wd_comp_setting.pool, msg->tag);
 
 	return ret;
 }