--- a/wd_ecc.c
+++ b/wd_ecc.c
@@ -63,7 +63,7 @@ static struct wd_ecc_setting {
 	enum wd_status status;
 	struct wd_ctx_config_internal config;
 	struct wd_sched sched;
-	struct wd_async_msg_pool pool;
+	struct msg_pool pool;
 	struct wd_alg_driver *driver;
 	void *dlhandle;
 	void *dlh_list;
@@ -2228,7 +2228,7 @@ int wd_do_ecc_async(handle_t sess, struct wd_ecc_req *req)
 
 	ctx = config->ctxs + idx;
 
-	mid = wd_get_msg_from_pool(&wd_ecc_setting.pool, idx, (void **)&msg);
+	mid = wd_get_msg_from_pool(&wd_ecc_setting.pool, (void **)&msg);
 	if (mid < 0)
 		return -WD_EBUSY;
 
@@ -2253,13 +2253,13 @@ int wd_do_ecc_async(handle_t sess, struct wd_ecc_req *req)
 	return 0;
 
 fail_with_msg:
-	wd_put_msg_to_pool(&wd_ecc_setting.pool, idx, mid);
+	wd_put_msg_to_pool(&wd_ecc_setting.pool, mid);
 	return ret;
 }
 
-struct wd_ecc_msg *wd_ecc_get_msg(__u32 idx, __u32 tag)
+struct wd_ecc_msg *wd_ecc_get_msg(__u32 tag)
 {
-	return wd_find_msg_in_pool(&wd_ecc_setting.pool, idx, tag);
+	return wd_find_msg_in_pool(&wd_ecc_setting.pool, tag);
 }
 
 int wd_ecc_poll_ctx(__u32 idx, __u32 expt, __u32 *count)
@@ -2292,13 +2292,11 @@ int wd_ecc_poll_ctx(__u32 idx, __u32 expt, __u32 *count)
 		} else if (ret < 0) {
 			WD_ERR("failed to async recv, ret = %d!\n", ret);
 			*count = rcv_cnt;
-			wd_put_msg_to_pool(&wd_ecc_setting.pool, idx,
-					   recv_msg.tag);
+			wd_put_msg_to_pool(&wd_ecc_setting.pool, recv_msg.tag);
 			return ret;
 		}
 		rcv_cnt++;
-		msg = wd_find_msg_in_pool(&wd_ecc_setting.pool, idx,
-					  recv_msg.tag);
+		msg = wd_find_msg_in_pool(&wd_ecc_setting.pool, recv_msg.tag);
 		if (!msg) {
 			WD_ERR("failed to get msg from pool!\n");
 			return -WD_EINVAL;
@@ -2308,7 +2306,7 @@ int wd_ecc_poll_ctx(__u32 idx, __u32 expt, __u32 *count)
 		msg->req.status = recv_msg.result;
 		req = &msg->req;
 		req->cb(req);
-		wd_put_msg_to_pool(&wd_ecc_setting.pool, idx, recv_msg.tag);
+		wd_put_msg_to_pool(&wd_ecc_setting.pool, recv_msg.tag);
 		*count = rcv_cnt;
 	} while (--tmp);