--- a/wd_rsa.c
+++ b/wd_rsa.c
@@ -71,7 +71,7 @@ static struct wd_rsa_setting {
 	enum wd_status status;
 	struct wd_ctx_config_internal config;
 	struct wd_sched sched;
-	struct wd_async_msg_pool pool;
+	struct msg_pool pool;
 	struct wd_alg_driver *driver;
 	void *dlhandle;
 	void *dlh_list;
@@ -451,7 +451,7 @@ int wd_do_rsa_async(handle_t sess, struct wd_rsa_req *req)
 
 	ctx = config->ctxs + idx;
 
-	mid = wd_get_msg_from_pool(&wd_rsa_setting.pool, idx, (void **)&msg);
+	mid = wd_get_msg_from_pool(&wd_rsa_setting.pool, (void **)&msg);
 	if (mid < 0)
 		return -WD_EBUSY;
 
@@ -476,13 +476,13 @@ int wd_do_rsa_async(handle_t sess, struct wd_rsa_req *req)
 	return 0;
 
 fail_with_msg:
-	wd_put_msg_to_pool(&wd_rsa_setting.pool, idx, mid);
+	wd_put_msg_to_pool(&wd_rsa_setting.pool, mid);
 	return ret;
 }
 
-struct wd_rsa_msg *wd_rsa_get_msg(__u32 idx, __u32 tag)
+struct wd_rsa_msg *wd_rsa_get_msg(__u32 tag)
 {
-	return wd_find_msg_in_pool(&wd_rsa_setting.pool, idx, tag);
+	return wd_find_msg_in_pool(&wd_rsa_setting.pool, tag);
 }
 
 int wd_rsa_poll_ctx(__u32 idx, __u32 expt, __u32 *count)
@@ -514,13 +514,11 @@ int wd_rsa_poll_ctx(__u32 idx, __u32 expt, __u32 *count)
 			return ret;
 		} else if (ret < 0) {
 			WD_ERR("failed to async recv, ret = %d!\n", ret);
-			wd_put_msg_to_pool(&wd_rsa_setting.pool, idx,
-					   recv_msg.tag);
+			wd_put_msg_to_pool(&wd_rsa_setting.pool, recv_msg.tag);
 			return ret;
 		}
 		rcv_cnt++;
-		msg = wd_find_msg_in_pool(&wd_rsa_setting.pool, idx,
-					  recv_msg.tag);
+		msg = wd_find_msg_in_pool(&wd_rsa_setting.pool, recv_msg.tag);
 		if (!msg) {
 			WD_ERR("failed to get msg from pool!\n");
 			return -WD_EINVAL;
@@ -530,7 +528,7 @@ int wd_rsa_poll_ctx(__u32 idx, __u32 expt, __u32 *count)
 		msg->req.status = recv_msg.result;
 		req = &msg->req;
 		req->cb(req);
-		wd_put_msg_to_pool(&wd_rsa_setting.pool, idx, recv_msg.tag);
+		wd_put_msg_to_pool(&wd_rsa_setting.pool, recv_msg.tag);
 		*count = rcv_cnt;
 	} while (--tmp);