The msg_pool does not need two dimensions, one dimension is enough, get the first empty msg from the list and use the index as tag. So the hardware info ctx_id can be removed.