@gaozhangfei 2023-11-29
for one algorithm, is WD_POOL_MAX_ENTRIES(1024 ) enough?
and user can set
/* If user set valid msg num, use user's. */
        get_ctx_msg_num(config->cap, &msg_num);

Now we alloc  async_ctx_num * WD_POOL_MAX_ENTRIES
async_ctx_num is async * nodes * op, is this a waste?