@Liulongfang 2023-12-07
> for one algorithm, is WD_POOL_MAX_ENTRIES(1024 ) msgs enough? and user can set msg num /* If user set valid msg num, use user's. */ get_ctx_msg_num(config->cap, &msg_num);
> 
> Now we alloc async_ctx_num * WD_POOL_MAX_ENTRIES async_ctx_num is async * nodes * op, is this a waste?

Now after modification, as a whole, performance bottlenecks may occur in multi-threaded scenarios. You can use uadk_tool benchmark to test the modified performance to ensure that performance is not affected.