@gaozhangfei 2023-12-07
> > for one algorithm, is WD_POOL_MAX_ENTRIES(1024 ) msgs enough? and user can set msg num /* If user set valid msg num, use user's. */ get_ctx_msg_num(config->cap, &msg_num);
> > Now we alloc async_ctx_num * WD_POOL_MAX_ENTRIES async_ctx_num is async * nodes * op, is this a waste?
> 
> Now after modification, as a whole, performance bottlenecks may occur in multi-threaded scenarios. You can use uadk_tool benchmark to test the modified performance to ensure that performance is not affected.

Yes, I am also concern the perf for multi-thread, but looks not see difference from the test
Maybe the incorrect test script?

with this patch
aes-128-ecb     8192Bytes       7893362.0KB/s   986.7Kops       799.55%
aes-128-ecb     8192Bytes       7895578.0KB/s   986.9Kops       799.60%
aes-128-ecb     8192Bytes       7895217.6KB/s   986.9Kops       799.60%
aes-128-ecb     8192Bytes       7899500.8KB/s   987.4Kops       799.65%
aes-128-ecb     8192Bytes       7898315.6KB/s   987.3Kops       799.50%
aes-128-ecb     8192Bytes       7904385.2KB/s   988.0Kops       799.60%

original
aes-128-ecb     8192Bytes       7897656.4KB/s   987.2Kops       799.55%
aes-128-ecb     8192Bytes       7896188.0KB/s   987.0Kops       799.55%
aes-128-ecb     8192Bytes       7896778.8KB/s   987.1Kops       799.60%


Test
for x in {1..3}
do
        echo $x
        numactl --cpubind=0 --membind=0 \
        ./uadk_tool benchmark --alg aes-128-ecb --mode sva --opt 0 --sync --pktlen 8192 --seconds 20 --multi 1 --thread 8 --ctxnum 8
done

The reason is we use ctx (hardware) info to get message in algorithm layer, which is a bit strange.
The msg should not related with hardware, only software related, ie, algo get one msg in send, and get such msg with idx in asyn recv, it does not matter which ctx.


For example, we may move the ctx, pool, shced out of the algorithm layer.