@Liulongfang 2023-12-11
> > > for one algorithm, is WD_POOL_MAX_ENTRIES(1024 ) msgs enough? and user can set msg num /* If user set valid msg num, use user's. */ get_ctx_msg_num(config->cap, &msg_num);
> > > Now we alloc async_ctx_num * WD_POOL_MAX_ENTRIES async_ctx_num is async * nodes * op, is this a waste?
> > 
> > 
> > Now after modification, as a whole, performance bottlenecks may occur in multi-threaded scenarios. You can use uadk_tool benchmark to test the modified performance to ensure that performance is not affected.
> 
> Yes, I am also concern the perf for multi-thread, but looks not see difference from the test Maybe the incorrect test script?
> 
> with this patch aes-128-ecb 8192Bytes 9089525.2KB/s 1136.2Kops 1483.80% aes-128-ecb 8192Bytes 9107025.6KB/s 1138.4Kops 1487.70% aes-128-ecb 8192Bytes 9093035.2KB/s 1136.6Kops 1490.65%
> 
> original aes-128-ecb 8192Bytes 9079088.4KB/s 1134.9Kops 1481.55% aes-128-ecb 8192Bytes 9115990.0KB/s 1139.5Kops 1486.25% aes-128-ecb 8192Bytes 9087795.6KB/s 1136.0Kops 1482.60%
> 
> Test for x in {1..3} do echo $x numactl --cpubind=0 --membind=0 ./uadk_tool benchmark --alg aes-128-ecb --mode sva --opt 0 --async --pktlen 8192 --seconds 20 --multi 1 --thread 8 --ctxnum 8 done
> 
> The reason is we use ctx (hardware) info to get message in algorithm layer, which is a bit strange. The msg should not related with hardware, only software related, ie, algo get one msg in send, and get such msg with idx in asyn recv, it does not matter which ctx.
> 
> For example, we may move the ctx, pool, shced out of the algorithm layer.

Your test mode above is an 8KB long packet in asynchronous mode. This kind of long packet is not sensitive to the time it takes to obtain the resource pool. In this case, you need to test the 1024Byte short message. And test 16, 24, 32, 48, 64 threads and multiple queues concurrently.
Another question, all your queues share a msg_pool. Then as the number of queues increases, the number of msg each thread can obtain will decrease. The competition between them will intensify, and the while operation in wd_get_msg_from_pool() will cause query failure.
However, you haven't encountered any of them here. Therefore, this test still cannot explain the problem.