@gaozhangfei 2023-12-11
有具体测试命令么，给个命令我测下看下对性能影响多大。
init2 现在只能用env? 参数指定多ctx没用。

我看其他msg pool也没这么按多线程设计的。就是给个size, num就完了。get_msg可能不是瓶颈。

主要是考虑到软件硬件共用一个pool, 
现在的方法绑定ctx, 除非软件也申请ctx, 且id 不同，不然有冲突。那就得统一申请ctx。
而且全耦合在一起了。算法层得知道硬件ctx 信息。
当然其实相同ctx id也没事，一个pool也行，

现在指令集也就申请2个ctx， 一个同步，一个异步。其实多线程也只用到了一维的pool, 对应那个异步ctx。

我也在考虑可有什么更好的办法，比如一个sess一个pool，
但poll 这里是统一根据tag来找msg，除非改动poll, 传递sess信息。