@Liulongfang 2023-12-12
> 这套设计软算都用不上啊，软算多线程用的都是一个ctx_id的pool 不可能为了用多维的pool, 故意申请多个ctx, 且每个ctx id都不一样吧
> 
> 即便是硬件，ctx 跟多线程也没必然联系吧，比如2个ctx, 8个线程啥的。
> 
> ctx是硬件，消息队列纯软件，这两个强耦合，真的好么。

软算现在是用不上的，但是需要保证硬算功能和性能正常。