@gaozhangfei 2023-12-13
测下来确实有性能损耗，
上次可能没测对，因为是库，需要重编